% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnishgrid.R
\name{afrr_activated_down}
\alias{afrr_activated_down}
\title{Automatic Frequency Restoration Reserve, activated, down}
\usage{
afrr_activated_down(start_time_utc = NA, end_time_utc = NA, user_key = NA)
}
\arguments{
\item{start_time_utc}{Start time in UTC. Character array YYYY-MM-ddTHH:mm:ss.sssZ}

\item{end_time_utc}{End time in UTC. Character array YYYY-MM-ddTHH:mm:ss.sssZ}

\item{user_key}{Character array holding API-key.
Free from https://data.fingrid.fi/en/instructions}
}
\value{
A data frame object with time series data having period 15 min and unit type MW.
}
\description{
The values in the data have been erroneously instantaneous values. From February 21, 2024, the data are correct 15-minute averages.Hourly values can be found correctly on the European transparency platform: https://transparency.entsoe.eu/balancing/r2/activationAndActivatedBalancingReserves/showActivated automatic Frequency Restoration Reserve (aFRR) energy, down (MW). Value is activated average power. The amount of activated Automatic Frequency. The Data before 13.06.2023 is in hourly resolution.
}
\examples{
\dontrun{
library(finnishgrid)
start = "2024-06-01T00:00:00.000Z"  # UTC
end = "2024-06-03T00:00:00.000Z"    # UTC
key = "MY_SUPER_SECRET"
df <- afrr_activated_down(start_time_utc = start,
         end_time_utc = end,
         user_key = key)
summary(df)
}
}
\seealso{
https://data.fingrid.fi/en/datasets/53
}
