% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnishgrid.R
\name{afrr_capacity_up}
\alias{afrr_capacity_up}
\title{Automatic Frequency Restoration Reserve, capacity, up}
\usage{
afrr_capacity_up(start_time_utc = NA, end_time_utc = NA, user_key = NA)
}
\arguments{
\item{start_time_utc}{Start time in UTC. Character array YYYY-MM-ddTHH:mm:ss.sssZ}

\item{end_time_utc}{End time in UTC. Character array YYYY-MM-ddTHH:mm:ss.sssZ}

\item{user_key}{Character array holding API-key.
Free from https://data.fingrid.fi/en/instructions}
}
\value{
A data frame object with time series data having period 1 h and unit type MW.
}
\description{
Procured automatic Frequency Restoration Reserve (aFRR) capacity, up (MW)
}
\examples{
\dontrun{
library(finnishgrid)
start = "2024-06-01T00:00:00.000Z"  # UTC
end = "2024-06-03T00:00:00.000Z"    # UTC
key = "MY_SUPER_SECRET"
df <- afrr_capacity_up(start_time_utc = start,
         end_time_utc = end,
         user_key = key)
summary(df)
}
}
\seealso{
https://data.fingrid.fi/en/datasets/1
}
