% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnishgrid.R
\name{day_ahead_transmission_capacity_FI_to_SE1}
\alias{day_ahead_transmission_capacity_FI_to_SE1}
\title{Day-ahead transmission capacity FI-SE1}
\usage{
day_ahead_transmission_capacity_FI_to_SE1(
  start_time = NA,
  end_time = NA,
  user_key = NA
)
}
\arguments{
\item{start_time}{Start time in UTC with offset. Character array in ISO8601,
YYYY-MM-ddTHH:mm:ssZ}

\item{end_time}{End time in UTC with offset. Character array in ISO8601,
YYYY-MM-ddTHH:mm:ssZ}

\item{user_key}{Character array holding API-key.
Free from https://data.fingrid.fi/open-data-forms/registration/}
}
\description{
Check https://data.fingrid.fi/en/dataset/vuorokausimarkkinoille-annettu-siirtokapasiteetti-fi-se1
}
