% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnishgrid.R
\name{fcr_n_yearlymarket_plans}
\alias{fcr_n_yearlymarket_plans}
\title{Frequency Containment Reserve for Normal operation, yearly market plans}
\usage{
fcr_n_yearlymarket_plans(start_time_utc = NA, end_time_utc = NA, user_key = NA)
}
\arguments{
\item{start_time_utc}{Start time in UTC. Character array YYYY-MM-ddTHH:mm:ss.sssZ}

\item{end_time_utc}{End time in UTC. Character array YYYY-MM-ddTHH:mm:ss.sssZ}

\item{user_key}{Character array holding API-key.
Free from https://data.fingrid.fi/en/instructions}
}
\value{
A data frame object with time series data having period 1 h and unit type MW.
}
\description{
The hourly sum of reserve plans for frequency containment reserve for normal operation (FCR-N) in the yearly market. The data will be published 22:45 (EET) on previous evening.FCR-N is the frequency containment reserve used in the Nordic synchronous system that aims to keep the frequency in normal frequency range between 49,9 - 50,1 Hz.Yearly market is a reserve market operated by Fingrid. Hourly procured volumes vary according to the reserve plans submitted by the balancing service providers and the price is constant over the whole year.
}
\examples{
\dontrun{
library(finnishgrid)
start = "2024-06-01T00:00:00.000Z"  # UTC
end = "2024-06-03T00:00:00.000Z"    # UTC
key = "MY_SUPER_SECRET"
df <- fcr_n_yearlymarket_plans(start_time_utc = start,
         end_time_utc = end,
         user_key = key)
summary(df)
}
}
\seealso{
https://data.fingrid.fi/en/datasets/288
}
