% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnishgrid.R
\name{imbalance_voaa_no4}
\alias{imbalance_voaa_no4}
\title{Value of Avoided Activation (VoAA) NO4}
\usage{
imbalance_voaa_no4(start_time_utc = NA, end_time_utc = NA, user_key = NA)
}
\arguments{
\item{start_time_utc}{Start time in UTC. Character array YYYY-MM-ddTHH:mm:ss.sssZ}

\item{end_time_utc}{End time in UTC. Character array YYYY-MM-ddTHH:mm:ss.sssZ}

\item{user_key}{Character array holding API-key.
Free from https://data.fingrid.fi/en/instructions}
}
\value{
A data frame object with time series data having period 1 h and unit type 1 €/MWh.
}
\description{
Value of Avoided Activation (VoAA) NO4More information about VoAA and IC can be found in eSett Handbook (definitions and calculation rules). VoAA and IC are not implemented in Norway as in the other Nordic countries, but corresponding values are published for informational purposes. In case VoAA and IC cannot be calculated according to the set rules, the value 99 999 will be displayed for both.
}
\examples{
\dontrun{
library(finnishgrid)
start = "2024-06-01T00:00:00.000Z"  # UTC
end = "2024-06-03T00:00:00.000Z"    # UTC
key = "MY_SUPER_SECRET"
df <- imbalance_voaa_no4(start_time_utc = start,
         end_time_utc = end,
         user_key = key)
summary(df)
}
}
\seealso{
https://data.fingrid.fi/en/datasets/297
}
