% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnishgrid.R
\name{ordered_up_regulations_balancing_energy_market_FI}
\alias{ordered_up_regulations_balancing_energy_market_FI}
\title{Ordered up-regulations from Balancing energy market in Finland}
\usage{
ordered_up_regulations_balancing_energy_market_FI(
  start_time = NA,
  end_time = NA,
  user_key = NA
)
}
\arguments{
\item{start_time}{Start time in UTC with offset. Character array in ISO8601,
YYYY-MM-ddTHH:mm:ssZ}

\item{end_time}{End time in UTC with offset. Character array in ISO8601,
YYYY-MM-ddTHH:mm:ssZ}

\item{user_key}{Character array holding API-key.
Free from https://data.fingrid.fi/open-data-forms/registration/}
}
\description{
Check https://data.fingrid.fi/en/dataset/ordered-up-regulations-from-balancing-energy-market-in-finland
}
