% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnishgrid.R
\name{other_production_RTD}
\alias{other_production_RTD}
\title{Other production inc. estimated small-scale production and reserve power plants - real time data}
\usage{
other_production_RTD(start_time = NA, end_time = NA, user_key = NA)
}
\arguments{
\item{start_time}{Start time in UTC with offset. Character array in ISO8601,
YYYY-MM-ddTHH:mm:ssZ}

\item{end_time}{End time in UTC with offset. Character array in ISO8601,
YYYY-MM-ddTHH:mm:ssZ}

\item{user_key}{Character array holding API-key.
Free from https://data.fingrid.fi/open-data-forms/registration/}
}
\description{
Check https://data.fingrid.fi/en/dataset/gas-turbine-and-estimated-small-scale-production-real-time-data
}
