% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnishgrid.R
\name{planned_day_ahead_transmission_capacity_SE3_to_FI}
\alias{planned_day_ahead_transmission_capacity_SE3_to_FI}
\title{Planned day-ahead transmission capacity SE3-FI}
\usage{
planned_day_ahead_transmission_capacity_SE3_to_FI(
  start_time = NA,
  end_time = NA,
  user_key = NA
)
}
\arguments{
\item{start_time}{Start time in UTC with offset. Character array in ISO8601,
YYYY-MM-ddTHH:mm:ssZ}

\item{end_time}{End time in UTC with offset. Character array in ISO8601,
YYYY-MM-ddTHH:mm:ssZ}

\item{user_key}{Character array holding API-key.
Free from https://data.fingrid.fi/open-data-forms/registration/}
}
\description{
Check https://data.fingrid.fi/en/dataset/planned-day-ahead-transmission-capacity-se3-fi
}
