% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnishgrid.R
\name{powersys_trans_FI_EE_RTD}
\alias{powersys_trans_FI_EE_RTD}
\title{Transmission between Finland and Estonia - real time data}
\usage{
powersys_trans_FI_EE_RTD(start_time_utc = NA, end_time_utc = NA, user_key = NA)
}
\arguments{
\item{start_time_utc}{Start time in UTC. Character array YYYY-MM-ddTHH:mm:ss.sssZ}

\item{end_time_utc}{End time in UTC. Character array YYYY-MM-ddTHH:mm:ss.sssZ}

\item{user_key}{Character array holding API-key.
Free from https://data.fingrid.fi/en/instructions}
}
\value{
A data frame object with time series data having period 3 min and unit type MW.
}
\description{
Power transmission between Finland and Estonia HVDC tie lines (Estlink 1 and Estlink 2). Data is based on the real-time measurements in Fingrid's operation control system. Positive sign means transmission from Finland to Estonia. Negative sign means transmission from Estonia to Finland. The data is updated every 3 minutes.
}
\examples{
\dontrun{
library(finnishgrid)
start = "2024-06-01T00:00:00.000Z"  # UTC
end = "2024-06-03T00:00:00.000Z"    # UTC
key = "MY_SUPER_SECRET"
df <- powersys_trans_FI_EE_RTD(start_time_utc = start,
         end_time_utc = end,
         user_key = key)
summary(df)
}
}
\seealso{
https://data.fingrid.fi/en/datasets/180
}
