% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnishgrid.R
\name{price_last_activated_down_regulation_bid_RTD}
\alias{price_last_activated_down_regulation_bid_RTD}
\title{Price of last activated down-regulation bid, real-time publication}
\usage{
price_last_activated_down_regulation_bid_RTD(
  start_time = NA,
  end_time = NA,
  user_key = NA
)
}
\arguments{
\item{start_time}{Start time in UTC with offset. Character array in ISO8601,
YYYY-MM-ddTHH:mm:ssZ}

\item{end_time}{End time in UTC with offset. Character array in ISO8601,
YYYY-MM-ddTHH:mm:ssZ}

\item{user_key}{Character array holding API-key.
Free from https://data.fingrid.fi/open-data-forms/registration/}
}
\description{
Check https://data.fingrid.fi/en/dataset/viimeksi-aktivoidun-alassaatotarjouksen-hinta-reaaliaikainen-julkaisu
}
