% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnishgrid.R
\name{production_surplus_or_deficit_FI_RTD}
\alias{production_surplus_or_deficit_FI_RTD}
\title{Production surplus/deficit in Finland - Real time}
\usage{
production_surplus_or_deficit_FI_RTD(
  start_time = NA,
  end_time = NA,
  user_key = NA
)
}
\arguments{
\item{start_time}{Start time in UTC with offset. Character array in ISO8601,
YYYY-MM-ddTHH:mm:ssZ}

\item{end_time}{End time in UTC with offset. Character array in ISO8601,
YYYY-MM-ddTHH:mm:ssZ}

\item{user_key}{Character array holding API-key.
Free from https://data.fingrid.fi/open-data-forms/registration/}
}
\description{
Check data	https://data.fingrid.fi/en/dataset/production-surplus-deficit-in-finland-real-time-data
}
