% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnishgrid.R
\name{sum_up_regulation_bids_balancing_energy_market}
\alias{sum_up_regulation_bids_balancing_energy_market}
\title{The sum of the up-regulation bids in the balancing energy market.}
\usage{
sum_up_regulation_bids_balancing_energy_market(
  start_time = NA,
  end_time = NA,
  user_key = NA
)
}
\arguments{
\item{start_time}{Start time in UTC with offset. Character array in ISO8601,
YYYY-MM-ddTHH:mm:ssZ}

\item{end_time}{End time in UTC with offset. Character array in ISO8601,
YYYY-MM-ddTHH:mm:ssZ}

\item{user_key}{Character array holding API-key.
Free from https://data.fingrid.fi/open-data-forms/registration/}
}
\description{
Check https://data.fingrid.fi/en/dataset/the-sum-of-the-up-regulation-bids-in-the-balancing-energy-market
}
