% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnishgrid.R
\name{transmission_capacity_intraday_market_SE1_to_FI}
\alias{transmission_capacity_intraday_market_SE1_to_FI}
\title{Transmission capacity for intraday market from Northern Sweden to Finland (SE1-FI)}
\usage{
transmission_capacity_intraday_market_SE1_to_FI(
  start_time = NA,
  end_time = NA,
  user_key = NA
)
}
\arguments{
\item{start_time}{Start time in UTC with offset. Character array in ISO8601,
YYYY-MM-ddTHH:mm:ssZ}

\item{end_time}{End time in UTC with offset. Character array in ISO8601,
YYYY-MM-ddTHH:mm:ssZ}

\item{user_key}{Character array holding API-key.
Free from https://data.fingrid.fi/open-data-forms/registration/}
}
\description{
Check https://data.fingrid.fi/en/dataset/transmission-capacity-to-be-given-to-intraday-market-se1-fi
}
