% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnishgrid.R
\name{transmission_capacity_RUS_to_FI}
\alias{transmission_capacity_RUS_to_FI}
\title{Transmission capacity RUS-FI}
\usage{
transmission_capacity_RUS_to_FI(start_time = NA, end_time = NA, user_key = NA)
}
\arguments{
\item{start_time}{Start time in UTC with offset. Character array in ISO8601,
YYYY-MM-ddTHH:mm:ssZ}

\item{end_time}{End time in UTC with offset. Character array in ISO8601,
YYYY-MM-ddTHH:mm:ssZ}

\item{user_key}{Character array holding API-key.
Free from https://data.fingrid.fi/open-data-forms/registration/}
}
\value{
A data frame object that contains the total commercial transmission
capacity of the 400 kV transmission lines from Russia to Finland owned
by Fingrid.
}
\description{
Check https://data.fingrid.fi/en/dataset/transmission-capacity-rus-fi
}
\examples{
\dontrun{
library(finnishgrid)
start = "2021-01-01T00:00:00+0200"  # UTC+2 offset, Helsinki time
end = "2021-01-10T00:00:00+0200"    # UTC+2 offset, Helsinki time
key = "MY_SUPER_SECRET"
df <- transmission_capacity_RUS_to_FI(start_time = start,
    end_time = end,
    user_key = key)
summary(df)
}
}
