% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnishgrid.R
\name{unused_biliteral_trade_capacity_RUS_to_FI}
\alias{unused_biliteral_trade_capacity_RUS_to_FI}
\title{Unused bilateral trade capacity RUS-FI}
\usage{
unused_biliteral_trade_capacity_RUS_to_FI(
  start_time = NA,
  end_time = NA,
  user_key = NA
)
}
\arguments{
\item{start_time}{Start time in UTC with offset. Character array in ISO8601,
YYYY-MM-ddTHH:mm:ssZ}

\item{end_time}{End time in UTC with offset. Character array in ISO8601,
YYYY-MM-ddTHH:mm:ssZ}

\item{user_key}{Character array holding API-key.
Free from https://data.fingrid.fi/open-data-forms/registration/}
}
\description{
Check https://data.fingrid.fi/en/dataset/unused-bilateral-trade-capacity-rus-fi
}
