% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnishgrid.R
\name{weekly_planned_capacity_P1_south_to_north}
\alias{weekly_planned_capacity_P1_south_to_north}
\title{Planned weekly capacity from south to north defined by Fingrid}
\usage{
weekly_planned_capacity_P1_south_to_north(
  start_time = NA,
  end_time = NA,
  user_key = NA
)
}
\arguments{
\item{start_time}{Start time in UTC with offset. Character array in ISO8601,
YYYY-MM-ddTHH:mm:ssZ}

\item{end_time}{End time in UTC with offset. Character array in ISO8601,
YYYY-MM-ddTHH:mm:ssZ}

\item{user_key}{Character array holding API-key.
Free from https://data.fingrid.fi/open-data-forms/registration/}
}
\description{
Check https://data.fingrid.fi/en/dataset/planned-weekly-capacity-from-south-to-north-defined-by-fingrid
}
