% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnishgrid.R
\name{wind_power_generation_forecast_daily_updated}
\alias{wind_power_generation_forecast_daily_updated}
\title{Wind power generation forecast - updated once a day}
\usage{
wind_power_generation_forecast_daily_updated(
  start_time = NA,
  end_time = NA,
  user_key = NA
)
}
\arguments{
\item{start_time}{Start time in UTC with offset. Character array in ISO8601,
YYYY-MM-ddTHH:mm:ssZ}

\item{end_time}{End time in UTC with offset. Character array in ISO8601,
YYYY-MM-ddTHH:mm:ssZ}

\item{user_key}{Character array holding API-key.
Free from https://data.fingrid.fi/open-data-forms/registration/}
}
\description{
Check https://data.fingrid.fi/en/dataset/wind-power-generation-forecast-updated-once-a-day
}
