% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finnishgrid.R
\name{peak_load_power_RTD}
\alias{peak_load_power_RTD}
\title{Peak load power - real time data}
\usage{
peak_load_power_RTD(start_time = NA, end_time = NA, user_key = NA)
}
\arguments{
\item{start_time}{Start time in UTC with offset. Character array in ISO8601,
YYYY-MM-ddTHH:mm:ssZ}

\item{end_time}{End time in UTC with offset. Character array in ISO8601,
YYYY-MM-ddTHH:mm:ssZ}

\item{user_key}{Character array holding API-key.
Free from https://data.fingrid.fi/open-data-forms/registration/}
}
\value{
A data frame object that contains activated peak load power based on
the real-time measurements in Fingrid's operation control system
including peak load reserve activations and trial runs during winter
period. The data is updated every 3 minutes.
}
\description{
Check https://data.fingrid.fi/en/dataset/peak-load-power-real-time-data
}
\examples{
\dontrun{
library(finnishgrid)
start = "2021-01-01T00:00:00+0200"  # UTC+2 offset, Helsinki time
end = "2021-01-10T00:00:00+0200"    # UTC+2 offset, Helsinki time
key = "MY_SUPER_SECRET"
df <- peak_load_power_RTD(
    start_time = start,
    end_time = end,
    user_key = key)
summary(df)
}
}
